map_service/requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.3
sqlalchemy~=2.0.41
psycopg2-binary~=2.9.10
python-dotenv~=1.1.0
pydantic~=2.11.5
python-multipart~=0.0.20
redis~=6.2.0
rq~=2.3.3